moved resume host zone
